many changes... fixed indexing problems with dydt in the derivs function pulled some code out from inside the ocean loop so it gets calculated (esp dAdt and dTdt, but also stuff needed for dTdt in the main notebook) fixed problem where PE['m0'] was being overwritten in derivs fcn fixed indexing problem with the J indexes in PE in the derivs fcn fixed line plots
atm co2 now updates in time! and so do land pools!
I think this means land pools are probably working, and will look ok when atm co2 does
but.. the values are wrong when FF emissions start
might be a unit problem
or a problem with dAdt in the derivs fcn
thought i would commit now to preserve this progress and keep working on atm CO2